example/requirements.txt
pillow==9.5.0
whitenoise==6.4.0
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.1
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0